+++ samples/endpoint/nextjs/app/loading.tsx
@@ -1,0 +1,13 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export default function Loading() {
+  return (
+    <div className="flex flex-col items-center justify-center h-96">
+      <h1 className="text-4xl animate-bounce">Loading ...</h1>
+    </div>
+  );
+}
